+++ sdk/src/androidTest/java/ly/count/android/sdk/ModuleCrashTests.java
@@ -16,6 +16,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
+import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.junit.After;
@@ -24,6 +25,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
+import org.mockito.internal.util.collections.Sets;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
@@ -118,6 +120,7 @@ public class ModuleCrashTests {
     public void provideCustomCrashSegment_DuringInit() {
         Countly countly = new Countly();
         CountlyConfig cConfig = (new CountlyConfig(TestUtils.getContext(), "appkey", "http://test.count.ly")).setDeviceId("1234").setLoggingEnabled(true).enableCrashReporting();
+        int[] arr = { 1, 2, 3, 4, 5 };
 
         Map<String, Object> segm = new HashMap<>();
         segm.put("aa", "dd");
@@ -127,7 +130,7 @@ public class ModuleCrashTests {
         segm.put("3", true);
         segm.put("4", 45.4f);
         segm.put("41", new Object());
-        segm.put("42", new int[] { 1, 2 });
+        segm.put("42", arr);
 
         cConfig.setCustomCrashSegment(segm);
 
@@ -140,6 +143,7 @@ public class ModuleCrashTests {
         segm2.put("2", 1234.55d);
         segm2.put("3", true);
         segm2.put("4", 45.4f);
+        segm2.put("42", arr);
 
         Assert.assertEquals(segm2, countly.moduleCrash.customCrashSegments);
     }
@@ -353,9 +357,10 @@ public class ModuleCrashTests {
      */
     @Test
     public void recordHandledException_globalCrashFilter() throws JSONException {
+        int[] arr = new int[] { 1, 2, 3, 4, 5 };
         CountlyConfig cConfig = TestUtils.createBaseConfig();
         cConfig.metricProviderOverride = mmp;
-        cConfig.crashes.setCustomCrashSegmentation(TestUtils.map("secret", "Minato", "int", Integer.MAX_VALUE, "double", Double.MAX_VALUE, "bool", true, "long", Long.MAX_VALUE, "float", 1.1, "object", new Object(), "array", new int[] { 1, 2 }));
+        cConfig.crashes.setCustomCrashSegmentation(TestUtils.map("secret", "Minato", "int", Integer.MAX_VALUE, "double", Double.MAX_VALUE, "bool", true, "long", Long.MAX_VALUE, "float", 1.1, "object", new Object(), "array", arr));
         cConfig.crashes.setGlobalCrashFilterCallback(crash -> {
             if (crash.getStackTrace().contains("Secret")) {
                 return true;
@@ -388,6 +393,7 @@ public class ModuleCrashTests {
                 "bool", true,
                 "float", 1.1,
                 "long", Long.MAX_VALUE,
+                "array", new JSONArray(arr),
                 "sphinx_no", 324), 11, TestUtils.map("secret", "Minato"), Collections.singletonList("_ram_total"));
     }
 
@@ -435,7 +441,7 @@ public class ModuleCrashTests {
             Assert.assertEquals(20, crash.getCrashMetrics().size());
 
             crash.getCrashMetrics().put("5", new Object());
-            crash.getCrashMetrics().put("6", new int[] { 1, 2 });
+            crash.getCrashMetrics().put("6", new int[][] { new int[] { 1, 2 }, new int[] { 3, 4 } });
             return false;
         });
         Countly countly = new Countly().init(cConfig);
@@ -463,9 +469,10 @@ public class ModuleCrashTests {
      */
     @Test
     public void recordHandledException_basic() throws JSONException {
+        int[] arr = new int[] { 1, 2, 3, 4, 5 };
         CountlyConfig cConfig = TestUtils.createBaseConfig();
         cConfig.metricProviderOverride = mmp;
-        cConfig.crashes.setCustomCrashSegmentation(TestUtils.map("secret", "Minato", "int", Integer.MAX_VALUE, "double", Double.MAX_VALUE, "bool", true, "long", Long.MAX_VALUE, "float", 1.1, "object", new Object(), "array", new int[] { 1, 2 }));
+        cConfig.crashes.setCustomCrashSegmentation(TestUtils.map("secret", "Minato", "int", Integer.MAX_VALUE, "double", Double.MAX_VALUE, "bool", true, "long", Long.MAX_VALUE, "float", 1.1, "object", new Object(), "array", arr));
         cConfig.crashes.setGlobalCrashFilterCallback(crash -> false);
         Countly countly = new Countly().init(cConfig);
 
@@ -482,6 +489,7 @@ public class ModuleCrashTests {
                 "bool", true,
                 "long", Long.MAX_VALUE,
                 "float", 1.1,
+                "array", new JSONArray(arr),
                 "sphinx_no", 324), 0, new HashMap<>(), new ArrayList<>());
     }
 
@@ -552,11 +560,12 @@ public class ModuleCrashTests {
      */
     @Test
     public void recordException_globalCrashFilter_eliminateUnsupportedTypesFromCrashMetrics() throws JSONException {
+        int[] arr = new int[] { 1, 2, 3, 4, 5 };
         CountlyConfig cConfig = TestUtils.createBaseConfig();
         cConfig.metricProviderOverride = mmp;
         cConfig.crashes.setGlobalCrashFilterCallback(crash -> {
             crash.getCrashMetrics().put("5", new Object());
-            crash.getCrashMetrics().put("6", new int[] { 1, 2 });
+            crash.getCrashMetrics().put("6", arr);
             crash.getCrashMetrics().put("7", "7");
             crash.getCrashMetrics().put("8", 8);
             crash.getCrashMetrics().put("9", 9.9d);
@@ -572,6 +581,7 @@ public class ModuleCrashTests {
         Exception exception = new Exception("Some message");
         countly.crashes().recordUnhandledException(exception);
         validateCrash(extractStackTrace(exception), "", true, false, new ConcurrentHashMap<>(), 2, TestUtils.map(
+            "6", new JSONArray(arr),
             "7", "7",
             "8", 8,
             "9", 9.9,
@@ -593,7 +603,7 @@ public class ModuleCrashTests {
         cConfig.metricProviderOverride = mmp;
         cConfig.crashes.setGlobalCrashFilterCallback(crash -> {
             crash.getCrashMetrics().put("5", new Object());
-            crash.getCrashMetrics().put("6", new int[] { 1, 2 });
+            crash.getCrashMetrics().put("6", new int[][] { new int[] { 1, 2 }, new int[] { 3, 4 } });
 
             return false;
         });
@@ -729,11 +739,12 @@ public class ModuleCrashTests {
      */
     @Test
     public void recordException_globalCrashFilter_invalidCustomSegmentations() throws JSONException {
+        int[] arr = new int[] { 1, 2, 3, 4, 5 };
         CountlyConfig cConfig = TestUtils.createBaseConfig();
         cConfig.metricProviderOverride = mmp;
         cConfig.crashes.setGlobalCrashFilterCallback(crash -> {
             crash.getCrashSegmentation().put("5", new Object());
-            crash.getCrashSegmentation().put("6", new int[] { 1, 2 });
+            crash.getCrashSegmentation().put("6", arr);
             crash.getCrashSegmentation().put("7", "7");
             return false;
         });
@@ -742,7 +753,7 @@ public class ModuleCrashTests {
 
         Exception exception = new Exception("secret");
         countly.crashes().recordUnhandledException(exception);
-        validateCrash(extractStackTrace(exception), "", true, false, TestUtils.map("7", "7"), 8, new ConcurrentHashMap<>(), new ArrayList<>());
+        validateCrash(extractStackTrace(exception), "", true, false, TestUtils.map("7", "7", "6", new JSONArray(arr)), 8, new ConcurrentHashMap<>(), new ArrayList<>());
     }
 
     /**
@@ -1083,10 +1094,11 @@ public class ModuleCrashTests {
      */
     @Test
     public void internalLimits_recordException_maxValueSize() throws JSONException {
+        int[] arr = new int[] { 1, 2, 3, 4, 5 };
         CountlyConfig cConfig = TestUtils.createBaseConfig();
         cConfig.metricProviderOverride = mmp;
         cConfig.sdkInternalLimits.setMaxValueSize(5);
-        cConfig.crashes.setCustomCrashSegmentation(TestUtils.map("test", "123456", "integer", Integer.MAX_VALUE, "arr", new int[] { 1, 2, 3, 4, 5 }, "double", Double.MAX_VALUE, "bool", true, "float", 1.1, "object", new Object()));
+        cConfig.crashes.setCustomCrashSegmentation(TestUtils.map("test", "123456", "integer", Integer.MAX_VALUE, "arr", arr, "double", Double.MAX_VALUE, "bool", true, "float", 1.1, "object", new Object()));
         Countly countly = new Countly().init(cConfig);
 
         countly.crashes().addCrashBreadcrumb("Surpass");
@@ -1095,7 +1107,8 @@ public class ModuleCrashTests {
         Exception exception = new Exception("Some message");
         countly.crashes().recordUnhandledException(exception, TestUtils.map("case", "o the great one", "have", "dinosaur", "int1", Integer.MIN_VALUE, "double1", Double.MIN_VALUE, "bool", false));
 
-        validateCrash(extractStackTrace(exception), "Surpa\nYourL\nRight\n", true, false, TestUtils.map("test", "12345", "integer", Integer.MAX_VALUE, "int1", Integer.MIN_VALUE, "double", Double.MAX_VALUE, "double1", Double.MIN_VALUE, "bool", false, "float", 1.1, "have", "dinos", "case", "o the"),
+        validateCrash(extractStackTrace(exception), "Surpa\nYourL\nRight\n", true, false,
+            TestUtils.map("arr", new JSONArray(arr), "test", "12345", "integer", Integer.MAX_VALUE, "int1", Integer.MIN_VALUE, "double", Double.MAX_VALUE, "double1", Double.MIN_VALUE, "bool", false, "float", 1.1, "have", "dinos", "case", "o the"),
             0, new HashMap<>(),
             new ArrayList<>());
     }
@@ -1261,5 +1274,195 @@ public class ModuleCrashTests {
         expectedStackTrace.append("\nHaburayaHa");
 
         validateCrash(expectedStackTrace.toString(), "", true, false, new HashMap<>(), 16, new HashMap<>(), new ArrayList<>());
+    }
+
+    /**
+     * "recordHandledException" with Array segmentations
+     * Validate that all primitive types arrays are successfully recorded
+     * And validate that Object arrays are not recorded
+     * But Generic type of Object array which its values are only primitive types are recorded
+     *
+     * @throws JSONException if the JSON is not valid
+     */
+    @Test
+    public void recordHandledException_validateSupportedArrays() throws JSONException {
+        int[] arr = { 1, 2, 3, 4, 5, 6, 7, 8, 9, 10 };
+        boolean[] arrB = { true, false, true, false, true, false, true, false, true, false };
+        String[] arrS = { "1", "2", "3", "4", "5", "6", "7", "8", "9", "10" };
+        long[] arrL = { Long.MAX_VALUE, Long.MIN_VALUE };
+        double[] arrD = { Double.MAX_VALUE, Double.MIN_VALUE };
+        Long[] arrLO = { Long.MAX_VALUE, Long.MIN_VALUE };
+        Double[] arrDO = { Double.MAX_VALUE, Double.MIN_VALUE };
+        Boolean[] arrBO = { Boolean.TRUE, Boolean.FALSE };
+        Integer[] arrIO = { Integer.MAX_VALUE, Integer.MIN_VALUE };
+        Object[] arrObj = { "1", 1, 1.1d, true, 1.1f, Long.MAX_VALUE };
+        Object[] arrObjStr = { "1", "1", "1.1d", "true", "1.1f", "Long.MAX_VALUE" };
+
+        CountlyConfig countlyConfig = TestUtils.createBaseConfig();
+        countlyConfig.metricProviderOverride = mmp;
+        Countly countly = new Countly().init(countlyConfig);
+
+        Map<String, Object> segmentation = TestUtils.map(
+            "arr", arr,
+            "arrB", arrB,
+            "arrS", arrS,
+            "arrL", arrL,
+            "arrD", arrD,
+            "arrLO", arrLO,
+            "arrDO", arrDO,
+            "arrBO", arrBO,
+            "arrIO", arrIO,
+            "arrObj", arrObj,
+            "arrObjStr", arrObjStr
+        );
+
+        Exception exception = new Exception("Some message");
+        countly.crashes().recordHandledException(exception, segmentation);
+
+        Map<String, Object> expectedSegmentation = TestUtils.map(
+            "arr", new JSONArray(arr),
+            "arrB", new JSONArray(arrB),
+            "arrS", new JSONArray(arrS),
+            "arrL", new JSONArray(arrL),
+            "arrD", new JSONArray(arrD),
+            "arrLO", new JSONArray(arrLO),
+            "arrDO", new JSONArray(arrDO),
+            "arrBO", new JSONArray(arrBO),
+            "arrIO", new JSONArray(arrIO)
+        );
+
+        validateCrash(extractStackTrace(exception), "", false, false, expectedSegmentation, 0, new HashMap<>(), new ArrayList<>());
+    }
+
+    /**
+     * "recordHandledException" with List segmentations
+     * Validate that all primitive types Lists are successfully recorded
+     * And validate that List of Objects is not recorded
+     * But Generic type of Object list which its values are only primitive types are recorded
+     *
+     * @throws JSONException if the JSON is not valid
+     */
+    @Test
+    public void recordHandledException_validateSupportedLists() throws JSONException {
+        List<Integer> arr = Arrays.asList(1, 2, 3, 4, 5, 6, 7, 8, 9, 10);
+        List<Boolean> arrB = Arrays.asList(true, false, true, false, true, false, true, false, true, false);
+        List<String> arrS = Arrays.asList("1", "2", "3", "4", "5", "6", "7", "8", "9", "10");
+        List<Long> arrLO = Arrays.asList(Long.MAX_VALUE, Long.MIN_VALUE);
+        List<Double> arrDO = Arrays.asList(Double.MAX_VALUE, Double.MIN_VALUE);
+        List<Boolean> arrBO = Arrays.asList(Boolean.TRUE, Boolean.FALSE);
+        List<Integer> arrIO = Arrays.asList(Integer.MAX_VALUE, Integer.MIN_VALUE);
+        List<Object> arrObj = Arrays.asList("1", 1, 1.1d, true, Long.MAX_VALUE);
+        List<Object> arrObjStr = Arrays.asList("1", "1", "1.1d", "true", "Long.MAX_VALUE");
+
+        CountlyConfig countlyConfig = TestUtils.createBaseConfig();
+        countlyConfig.metricProviderOverride = mmp;
+        Countly countly = new Countly().init(countlyConfig);
+
+        // Create segmentation using maps with lists
+        Map<String, Object> segmentation = TestUtils.map(
+            "arr", arr,
+            "arrB", arrB,
+            "arrS", arrS,
+            "arrLO", arrLO,
+            "arrDO", arrDO,
+            "arrBO", arrBO,
+            "arrIO", arrIO,
+            "arrObj", arrObj,
+            "arrObjStr", arrObjStr
+        );
+
+        Exception exception = new Exception("Some message");
+        countly.crashes().recordHandledException(exception, segmentation);
+
+        // Prepare expected segmentation with JSONArrays
+        Map<String, Object> expectedSegmentation = TestUtils.map(
+            "arr", new JSONArray(arr),
+            "arrB", new JSONArray(arrB),
+            "arrS", new JSONArray(arrS),
+            "arrLO", new JSONArray(arrLO),
+            "arrDO", new JSONArray(arrDO),
+            "arrBO", new JSONArray(arrBO),
+            "arrIO", new JSONArray(arrIO),
+            "arrObjStr", new JSONArray(arrObjStr)
+        );
+
+        validateCrash(extractStackTrace(exception), "", false, false, expectedSegmentation, 0, new HashMap<>(), new ArrayList<>());
+    }
+
+    /**
+     * "recordHandledException" with JSONArray segmentations
+     * Validate that all primitive types JSONArrays are successfully recorded
+     * And validate and JSONArray of Objects is not recorded
+     *
+     * @throws JSONException if the JSON is not valid
+     */
+    @Test
+    public void recordHandledException_validateSupportedJSONArrays() throws JSONException {
+        JSONArray arr = new JSONArray(Arrays.asList(1, 2, 3, 4, 5, 6, 7, 8, 9, 10));
+        JSONArray arrB = new JSONArray(Arrays.asList(true, false, true, false, true, false, true, false, true, false));
+        JSONArray arrS = new JSONArray(Arrays.asList("1", "2", "3", "4", "5", "6", "7", "8", "9", "10"));
+        JSONArray arrL = new JSONArray(Arrays.asList(Long.MAX_VALUE, Long.MIN_VALUE));
+        JSONArray arrD = new JSONArray(Arrays.asList(Double.MAX_VALUE, Double.MIN_VALUE));
+        JSONArray arrBO = new JSONArray(Arrays.asList(Boolean.TRUE, Boolean.FALSE));
+        JSONArray arrIO = new JSONArray(Arrays.asList(Integer.MAX_VALUE, Integer.MIN_VALUE));
+        JSONArray arrObj = new JSONArray(Arrays.asList("1", 1, 1.1d, true, Long.MAX_VALUE));
+
+        CountlyConfig countlyConfig = TestUtils.createBaseConfig();
+        countlyConfig.metricProviderOverride = mmp;
+        Countly countly = new Countly().init(countlyConfig);
+
+        // Create segmentation using maps with lists
+        Map<String, Object> segmentation = TestUtils.map(
+            "arr", arr,
+            "arrB", arrB,
+            "arrS", arrS,
+            "arrL", arrL,
+            "arrD", arrD,
+            "arrBO", arrBO,
+            "arrIO", arrIO,
+            "arrObj", arrObj
+        );
+
+        Exception exception = new Exception("Some message");
+        countly.crashes().recordHandledException(exception, segmentation);
+
+        // Prepare expected segmentation with JSONArrays
+        Map<String, Object> expectedSegmentation = TestUtils.map(
+            "arr", arr,
+            "arrB", arrB,
+            "arrS", arrS,
+            "arrL", arrL,
+            "arrD", arrD,
+            "arrBO", arrBO,
+            "arrIO", arrIO
+        );
+
+        validateCrash(extractStackTrace(exception), "", false, false, expectedSegmentation, 0, new HashMap<>(), new ArrayList<>());
+    }
+
+    /**
+     * "recordHandledException" with invalid data types
+     * Validate that unsupported data types are not recorded
+     *
+     * @throws JSONException if the JSON is not valid
+     */
+    @Test
+    public void recordHandledException_unsupportedDataTypesSegmentation() throws JSONException {
+        CountlyConfig countlyConfig = TestUtils.createBaseConfig();
+        countlyConfig.metricProviderOverride = mmp;
+        Countly countly = new Countly().init(countlyConfig);
+
+        Map<String, Object> segmentation = TestUtils.map(
+            "a", TestUtils.map(),
+            "b", TestUtils.json(),
+            "c", new Object(),
+            "d", Sets.newSet(),
+            "e", mmp
+        );
+
+        Exception exception = new Exception("Some message");
+        countly.crashes().recordHandledException(exception, segmentation);
+
+        validateCrash(extractStackTrace(exception), "", false, false, TestUtils.map(), 0, new HashMap<>(), new ArrayList<>());
     }
 }