+++ sdk/src/androidTest/java/ly/count/android/sdk/ModuleRemoteConfigTests.java
@@ -30,7 +30,7 @@ public class ModuleRemoteConfigTests {
     }
 
     /**
-     * Consent removal should clear stored remote config values
+     * Consent removal shouldn't clear stored remote config values
      */
     @Test
     public void valuesClearedOnConsentRemoval() {
@@ -50,7 +50,8 @@ public class ModuleRemoteConfigTests {
 
         countly.consent().removeConsentAll();
 
-        Assert.assertEquals(0, countly.remoteConfig().getValues().size());
+        Assert.assertEquals(2, countly.remoteConfig().getValues().size()); // values are cache cleared
+        countly.remoteConfig().getValues().forEach((k, v) -> Assert.assertTrue(v.isCurrentUsersData));
     }
 
     /**