+++ CHANGELOG.md
@@ -5,6 +5,14 @@
   * Upon flushing the event queue
  
 * Mitigated an issue where remote config values caching was changing by device id change, not anymore. It is linked to the configuration "enableRemoteConfigValueCaching" now.
+* Mitigated an issue related to the device ID by creating an internal migration.
+* Resolved a problem where revoked consents were sent after changes without merging.
+* Fixed a bug that caused the device ID to be incorrectly set after changes with merging.
+* Mitigated an issue where on consent revoke, remote config values were cleared, not anymore.
+
+* Added support for array, List and JSONArray to all user given segmentations. They will support only mutable and ummutable versions of the primitive types. Which are:
+  * String, Integer, int, Boolean, bool, Float, float, Double, double, Long, long
+  * Keep in mind that float array will be converted to the double array by the JSONArray
 
 ## 24.4.1
 * ! Minor breaking change ! Mitigated an issue where internal SDK limits did not apply